⚡️ improve : 상세페이지 댓글 컴포넌트화화
